Lift Notes browse molecules into the design system
